backend-api/pyproject.toml
[project]
name = "backend-api"
version = "1.0.0"
description = "Backend API for Embedding Recommender SaaS"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.109.0",
    "uvicorn[standard]==0.27.0",
    "sqlalchemy==2.0.25",
    "alembic==1.13.1",
    "psycopg2-binary==2.9.9",
    "asyncpg==0.29.0",
    "redis==5.0.1",
    "python-jose[cryptography]==3.3.0",
    "passlib[bcrypt]==1.7.4",
    "python-multipart==0.0.6",
    "pydantic==2.5.3",
    "pydantic-settings==2.1.0",
    "httpx==0.26.0",
    "prometheus-client==0.19.0",
    "python-json-logger==2.0.7",
    "structlog==24.1.0",
    "python-dotenv==1.0.0",
    "bcrypt==4.1.2",
]

[project.optional-dependencies]
dev = [
    "pytest==7.4.4",
    "pytest-asyncio==0.23.3",
    "pytest-cov==4.1.0",
    "black==24.1.1",
    "flake8==7.0.0",
    "mypy==1.8.0",
    "isort==5.13.2",
    "faker==22.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
